+++ dbt/include/spark/macros/materializations/incremental/incremental.sql
@@ -41,6 +41,8 @@
     {{ build_sql }}
   {%- endcall -%}
 
+  {% do persist_docs(target_relation, model) %}
+  
   {{ run_hooks(post_hooks) }}
 
   {{ return({'relations': [target_relation]}) }}